Рефакторинг доступа и добавление управления проектами
- Добавлен `IUserContextService` для работы с контекстом пользователя.
- Изменены `ClassController` и `InviteController` для проверки роли и идентификатора пользователя.
- Добавлен `ProjectController` для управления проектами.
- Введена новая модель `WorkProject` вместо `TaskProject`.
- Добавлены DTO и ViewModel для проектов.
- Обновлены репозитории и сервисы для унификации проверки доступа.
- Добавлено поле `Color` в `TopicClass` и связанные ViewModel.
- Обновлены маршруты API для соответствия RESTful-стилю.
- Удален устаревший `TaskProject`.
- Исправлены исключения в `AuthService` для более точной обработки ошибок.

+++ Services/UserService.cs
@@ -22,7 +22,7 @@ namespace GanttChartAPI.Services
                 Id = u.Id,
                 FullName = u.FullName,
                 Email = u.Email,
-                Role = u.Role,
+                Role = (int)u.Role,
             }).ToList();
         }
         public async Task<List<ClassRoleViewModel>> GetUserClassRolesAsync(Guid userId)
@@ -49,7 +49,7 @@ namespace GanttChartAPI.Services
                 Id = user.Id,
                 FullName = user.FullName,
                 Email = user.Email,
-                Role = user.Role
+                Role = (int)user.Role
             };
         }
     }
